removed float thing

+++ cyclone_game/tell_me_you_see_green.py
@@ -17,7 +17,7 @@ time.sleep(0.1)
 cap = cv2.VideoCapture(0)
 while True:
 	_, frame = cap.read()
-	hsv = cv2.cvtColor(np.float32(frame), cv2.COLOR_BGR2HSV)
+	hsv = cv2.cvtColor(frame, cv2.COLOR_BGR2HSV)
 
 	lower_green = np.array([0, 0, 0])
 	upper_green = np.array([255, 255, 255])
